[RTE] Fix issue when floating toolbar uses previous range for target reference
* Extracted `getVirtualReferenceElement` to separate helper file (`PositionedToolbar.helpers.ts`)
* Improved virtual element positioning with `useEffect` hook and `useForceUpdate`
* Added `useRef` to manage virtual target reference and ensure proper updates
* Updated floating-ui middleware: added `inline()` for better text selection positioning
